# Batch check improvements