--- a/internal/dispatch/caching/caching.go
+++ b/internal/dispatch/caching/caching.go
@@ -263,10 +263,17 @@ func (cd *Dispatcher) DispatchCheck(ctx context.Context, req *v1.DispatchCheckRe
 
 			// If debugging is requested, clone and add the req and the response to the trace.
 			clone := cachedResult.response.CloneVT()
+			results := make(map[string]*v1.CheckDebugTrace_ResourceCheckResult, len(cachedResult.response.ResultsByResourceId))
+			for resourceId, result := range cachedResult.response.ResultsByResourceId {
+				results[resourceId] = &v1.CheckDebugTrace_ResourceCheckResult{
+					HasPermission: result.Membership == v1.DispatchCheckResponse_MEMBER,
+				}
+			}
+
 			clone.Metadata.DebugInfo = &v1.DebugInformation{
 				Check: &v1.CheckDebugTrace{
 					Request:        req,
-					HasPermission:  clone.Membership == v1.DispatchCheckResponse_MEMBER,
+					Results:        results,
 					IsCachedResult: true,
 				},
 			}