--- a/internal/dispatch/dispatch.go
+++ b/internal/dispatch/dispatch.go
@@ -111,18 +111,35 @@ var cachePrefixes = []cachePrefix{checkViaRelationPrefix, checkViaCanonicalPrefi
 
 // CheckRequestToKey converts a check request into a cache key based on the relation
 func CheckRequestToKey(req *v1.DispatchCheckRequest) string {
-	return fmt.Sprintf("%s//%s@%s@%s", checkViaRelationPrefix, tuple.StringONR(req.ResourceAndRelation), tuple.StringONR(req.Subject), req.Metadata.AtRevision)
+	return fmt.Sprintf(
+		"%s//%s@%s:%s@%s:%d",
+		checkViaRelationPrefix,
+		tuple.StringRR(req.ResourceRelation),
+		strings.Join(req.ResourceIds, ","),
+		tuple.StringONR(req.Subject),
+		req.Metadata.AtRevision,
+		req.ResultsSetting,
+	)
 }
 
 // CheckRequestToKeyWithCanonical converts a check request into a cache key based
 // on the canonical key.
 func CheckRequestToKeyWithCanonical(req *v1.DispatchCheckRequest, canonicalKey string) string {
 	if canonicalKey == "" {
-		panic(fmt.Sprintf("given empty canonical key for request: %s => %s", req.ResourceAndRelation, tuple.StringONR(req.Subject)))
+		panic(fmt.Sprintf("given empty canonical key for request: %s => %s", req.ResourceRelation, tuple.StringONR(req.Subject)))
 	}
 
 	// NOTE: canonical cache keys are only unique *within* a version of a namespace.
-	return fmt.Sprintf("%s//%s:%s#%s@%s@%s", checkViaCanonicalPrefix, req.ResourceAndRelation.Namespace, req.ResourceAndRelation.ObjectId, canonicalKey, tuple.StringONR(req.Subject), req.Metadata.AtRevision)
+	return fmt.Sprintf(
+		"%s//%s:%s#%s@%s@%s:%d",
+		checkViaCanonicalPrefix,
+		req.ResourceRelation.Namespace,
+		strings.Join(req.ResourceIds, ","),
+		canonicalKey,
+		tuple.StringONR(req.Subject),
+		req.Metadata.AtRevision,
+		req.ResultsSetting,
+	)
 }
 
 // LookupRequestToKey converts a lookup request into a cache key