--- a/internal/dispatch/dispatch.go
+++ b/internal/dispatch/dispatch.go
@@ -4,14 +4,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"strings"
 
-	"github.com/jzelinskie/stringz"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 
 	v1 "github.com/authzed/spicedb/pkg/proto/dispatch/v1"
-	"github.com/authzed/spicedb/pkg/tuple"
 )
 
 // ErrMaxDepth is returned from CheckDepth when the max depth is exceeded.
@@ -94,94 +91,6 @@ func CheckDepth(ctx context.Context, req HasMetadata) error {
 	}
 
 	return nil
-}
-
-type cachePrefix string
-
-const (
-	checkViaRelationPrefix   cachePrefix = "cr"
-	checkViaCanonicalPrefix  cachePrefix = "cc"
-	lookupPrefix             cachePrefix = "l"
-	expandPrefix             cachePrefix = "e"
-	reachableResourcesPrefix cachePrefix = "rr"
-	lookupSubjectsPrefix     cachePrefix = "ls"
-)
-
-var cachePrefixes = []cachePrefix{checkViaRelationPrefix, checkViaCanonicalPrefix, lookupPrefix, expandPrefix, reachableResourcesPrefix}
-
-// CheckRequestToKey converts a check request into a cache key based on the relation
-func CheckRequestToKey(req *v1.DispatchCheckRequest) string {
-	return fmt.Sprintf(
-		"%s//%s@%s:%s@%s:%d",
-		checkViaRelationPrefix,
-		tuple.StringRR(req.ResourceRelation),
-		strings.Join(req.ResourceIds, ","),
-		tuple.StringONR(req.Subject),
-		req.Metadata.AtRevision,
-		req.ResultsSetting,
-	)
-}
-
-// CheckRequestToKeyWithCanonical converts a check request into a cache key based
-// on the canonical key.
-func CheckRequestToKeyWithCanonical(req *v1.DispatchCheckRequest, canonicalKey string) string {
-	if canonicalKey == "" {
-		panic(fmt.Sprintf("given empty canonical key for request: %s => %s", req.ResourceRelation, tuple.StringONR(req.Subject)))
-	}
-
-	// NOTE: canonical cache keys are only unique *within* a version of a namespace.
-	return fmt.Sprintf(
-		"%s//%s:%s#%s@%s@%s:%d",
-		checkViaCanonicalPrefix,
-		req.ResourceRelation.Namespace,
-		strings.Join(req.ResourceIds, ","),
-		canonicalKey,
-		tuple.StringONR(req.Subject),
-		req.Metadata.AtRevision,
-		req.ResultsSetting,
-	)
-}
-
-// LookupRequestToKey converts a lookup request into a cache key
-func LookupRequestToKey(req *v1.DispatchLookupRequest) string {
-	return fmt.Sprintf("%s//%s#%s@%s@%s", lookupPrefix, req.ObjectRelation.Namespace, req.ObjectRelation.Relation, tuple.StringONR(req.Subject), req.Metadata.AtRevision)
-}
-
-// ExpandRequestToKey converts an expand request into a cache key
-func ExpandRequestToKey(req *v1.DispatchExpandRequest) string {
-	return fmt.Sprintf("%s//%s@%s", expandPrefix, tuple.StringONR(req.ResourceAndRelation), req.Metadata.AtRevision)
-}
-
-// ReachableResourcesRequestToKey converts a reachable resources request into a cache key
-func ReachableResourcesRequestToKey(req *v1.DispatchReachableResourcesRequest) string {
-	return fmt.Sprintf("%s//%s#%s@%s#%s:[%s]@%s",
-		reachableResourcesPrefix,
-		req.ResourceRelation.Namespace,
-		req.ResourceRelation.Relation,
-		req.SubjectRelation.Namespace,
-		req.SubjectRelation.Relation,
-		strings.Join(req.SubjectIds, ","),
-		req.Metadata.AtRevision,
-	)
-}
-
-// LookupSubjectsRequestToKey converts a lookup subjects request into a cache key
-func LookupSubjectsRequestToKey(req *v1.DispatchLookupSubjectsRequest) string {
-	return stringz.Join("",
-		string(lookupSubjectsPrefix),
-		"//",
-		req.ResourceRelation.Namespace,
-		"#",
-		req.ResourceRelation.Relation,
-		"@",
-		req.SubjectRelation.Namespace,
-		"#",
-		req.SubjectRelation.Relation,
-		":",
-		strings.Join(req.ResourceIds, ","),
-		"@",
-		req.Metadata.AtRevision,
-	)
 }
 
 // AddResponseMetadata adds the metadata found in the incoming metadata to the existing