--- a/internal/dispatch/keys/computed.go
+++ b/internal/dispatch/keys/computed.go
@@ -30,9 +30,9 @@ var cachePrefixes = []cachePrefix{
 	lookupSubjectsPrefix,
 }
 
-// CheckRequestToKey converts a check request into a cache key based on the relation
-func CheckRequestToKey(req *v1.DispatchCheckRequest) DispatchCacheKey {
-	return dispatchCacheKeyHash(checkViaRelationPrefix, req.Metadata.AtRevision,
+// checkRequestToKey converts a check request into a cache key based on the relation
+func checkRequestToKey(req *v1.DispatchCheckRequest, option dispatchCacheKeyHashComputeOption) DispatchCacheKey {
+	return dispatchCacheKeyHash(checkViaRelationPrefix, req.Metadata.AtRevision, option,
 		hashableRelationReference{req.ResourceRelation},
 		hashableIds(req.ResourceIds),
 		hashableOnr{req.Subject},
@@ -40,15 +40,15 @@ func CheckRequestToKey(req *v1.DispatchCheckRequest) DispatchCacheKey {
 	)
 }
 
-// CheckRequestToKeyWithCanonical converts a check request into a cache key based
+// checkRequestToKeyWithCanonical converts a check request into a cache key based
 // on the canonical key.
-func CheckRequestToKeyWithCanonical(req *v1.DispatchCheckRequest, canonicalKey string) DispatchCacheKey {
+func checkRequestToKeyWithCanonical(req *v1.DispatchCheckRequest, canonicalKey string) DispatchCacheKey {
 	if canonicalKey == "" {
 		panic(fmt.Sprintf("given empty canonical key for request: %s => %s", req.ResourceRelation, tuple.StringONR(req.Subject)))
 	}
 
 	// NOTE: canonical cache keys are only unique *within* a version of a namespace.
-	return dispatchCacheKeyHash(checkViaCanonicalPrefix, req.Metadata.AtRevision,
+	return dispatchCacheKeyHash(checkViaCanonicalPrefix, req.Metadata.AtRevision, computeBothHashes,
 		hashableString(req.ResourceRelation.Namespace),
 		hashableString(canonicalKey),
 		hashableIds(req.ResourceIds),
@@ -57,33 +57,33 @@ func CheckRequestToKeyWithCanonical(req *v1.DispatchCheckRequest, canonicalKey s
 	)
 }
 
-// LookupRequestToKey converts a lookup request into a cache key
-func LookupRequestToKey(req *v1.DispatchLookupRequest) DispatchCacheKey {
-	return dispatchCacheKeyHash(lookupPrefix, req.Metadata.AtRevision,
+// lookupRequestToKey converts a lookup request into a cache key
+func lookupRequestToKey(req *v1.DispatchLookupRequest, option dispatchCacheKeyHashComputeOption) DispatchCacheKey {
+	return dispatchCacheKeyHash(lookupPrefix, req.Metadata.AtRevision, option,
 		hashableRelationReference{req.ObjectRelation},
 		hashableOnr{req.Subject},
 	)
 }
 
-// ExpandRequestToKey converts an expand request into a cache key
-func ExpandRequestToKey(req *v1.DispatchExpandRequest) DispatchCacheKey {
-	return dispatchCacheKeyHash(expandPrefix, req.Metadata.AtRevision,
+// expandRequestToKey converts an expand request into a cache key
+func expandRequestToKey(req *v1.DispatchExpandRequest, option dispatchCacheKeyHashComputeOption) DispatchCacheKey {
+	return dispatchCacheKeyHash(expandPrefix, req.Metadata.AtRevision, option,
 		hashableOnr{req.ResourceAndRelation},
 	)
 }
 
-// ReachableResourcesRequestToKey converts a reachable resources request into a cache key
-func ReachableResourcesRequestToKey(req *v1.DispatchReachableResourcesRequest) DispatchCacheKey {
-	return dispatchCacheKeyHash(reachableResourcesPrefix, req.Metadata.AtRevision,
+// reachableResourcesRequestToKey converts a reachable resources request into a cache key
+func reachableResourcesRequestToKey(req *v1.DispatchReachableResourcesRequest, option dispatchCacheKeyHashComputeOption) DispatchCacheKey {
+	return dispatchCacheKeyHash(reachableResourcesPrefix, req.Metadata.AtRevision, option,
 		hashableRelationReference{req.ResourceRelation},
 		hashableRelationReference{req.SubjectRelation},
 		hashableIds(req.SubjectIds),
 	)
 }
 
-// LookupSubjectsRequestToKey converts a lookup subjects request into a cache key
-func LookupSubjectsRequestToKey(req *v1.DispatchLookupSubjectsRequest) DispatchCacheKey {
-	return dispatchCacheKeyHash(lookupSubjectsPrefix, req.Metadata.AtRevision,
+// lookupSubjectsRequestToKey converts a lookup subjects request into a cache key
+func lookupSubjectsRequestToKey(req *v1.DispatchLookupSubjectsRequest, option dispatchCacheKeyHashComputeOption) DispatchCacheKey {
+	return dispatchCacheKeyHash(lookupSubjectsPrefix, req.Metadata.AtRevision, option,
 		hashableRelationReference{req.ResourceRelation},
 		hashableRelationReference{req.SubjectRelation},
 		hashableIds(req.ResourceIds),