--- a/internal/dispatch/keys/hasher_wasm.go
+++ b/internal/dispatch/keys/hasher_wasm.go
@@ -1,0 +1,5 @@
+package keys
+
+func dispatchCacheKeyHash(prefix cachePrefix, atRevision string, args ...hashableValue) DispatchCacheKey {
+	panic("Caching is not implemented under WASM")
+}