… internal/dispatch/caching/cachingdispatch_test.go
… internal/dispatch/dispatch_test.go
… internal/dispatch/graph/check_test.go
… internal/dispatch/graph/lookupsubjects_test.go
… internal/dispatch/keys/computed_test.go
… internal/dispatch/keys/hasher_common.go
… internal/dispatch/keys/hasher_ristretto.go
… internal/dispatch/keys/keys.go
… internal/dispatch/remote/cluster.go
… internal/graph/check.go
… internal/graph/check_test.go
… internal/graph/graph.go
… internal/graph/parallelchecker.go
… internal/services/benchconfigs/checkwidegroups.yaml
… internal/services/benchmark_test.go
… internal/services/testconfigs/directgroups.yaml
… internal/services/testconfigs/groupsintersection.yaml
… internal/services/testconfigs/indirectgroups.yaml
… internal/services/testconfigs/indirectnestedgroups.yaml
… internal/services/testconfigs/linuxfoundation.yaml
… internal/services/testconfigs/quay.yaml
… internal/services/v1/permissions.go
… internal/util/multimap.go
… pkg/cache/cache_ristretto.go
… pkg/development/check.go
… pkg/proto/dispatch/v1/00_zerolog.go
… pkg/proto/dispatch/v1/dispatch.pb.go
… pkg/proto/dispatch/v1/dispatch.pb.validate.go
… pkg/proto/dispatch/v1/dispatch_vtproto.pb.go
… pkg/tuple/onr.go