Implements support for batching of check requests to reduce latency for wide dispatches

Performance comparison:

```
name                                                          old time/op  new time/op
Services/basic_check_for_a_user/cockroachdb/v1-10             26.0ms       20.9ms   
Services/basic_check_for_a_user/mysql/v1-10                   6.00ms       5.44ms   
Services/basic_check_for_a_user/postgres/v1-10                6.85ms       4.79ms   
Services/basic_check_for_a_user/spanner/v1-10                 6.68ms       5.76ms   

Services/recursive_check_for_a_user/cockroachdb/v1-10         61.9ms       61.5ms   
Services/recursive_check_for_a_user/mysql/v1-10               13.3ms       13.6ms   
Services/recursive_check_for_a_user/postgres/v1-10            11.9ms       11.1ms   
Services/recursive_check_for_a_user/spanner/v1-10             29.0ms       20.7ms   

Services/wide_groups_check_for_a_user/cockroachdb/v1-10        274ms         73ms   
Services/wide_groups_check_for_a_user/mysql/v1-10             99.2ms       32.6ms   
Services/wide_groups_check_for_a_user/postgres/v1-10          66.3ms       13.0ms   
Services/wide_groups_check_for_a_user/spanner/v1-10            714ms        230ms   

name                                                           old alloc/op   new alloc/op
Services/basic_check_for_a_user/cockroachdb/v1-10              407kB          202kB     
Services/basic_check_for_a_user/mysql/v1-10                    151kB          148kB     
Services/basic_check_for_a_user/postgres/v1-10                 156kB          153kB     
Services/basic_check_for_a_user/spanner/v1-10                  209kB          207kB     

Services/recursive_check_for_a_user/cockroachdb/v1-10          703kB          681kB     
Services/recursive_check_for_a_user/mysql/v1-10                607kB          381kB     
Services/recursive_check_for_a_user/postgres/v1-10             436kB          836kB     
Services/recursive_check_for_a_user/spanner/v1-10              557kB          547kB     

Services/wide_groups_check_for_a_user/cockroachdb/v1-10       7.30MB         0.96MB     
Services/wide_groups_check_for_a_user/mysql/v1-10             6.40MB         0.47MB     
Services/wide_groups_check_for_a_user/postgres/v1-10          7.36MB         0.49MB     
Services/wide_groups_check_for_a_user/spanner/v1-10           11.6MB          1.4MB

name                                                           old allocs/op  new allocs/op
Services/basic_check_for_a_user/cockroachdb/v1-10              3.10k          3.00k          
Services/basic_check_for_a_user/mysql/v1-10                    2.57k          2.47k          
Services/basic_check_for_a_user/postgres/v1-10                 2.64k          2.54k          
Services/basic_check_for_a_user/spanner/v1-10                  3.15k          3.07k          

Services/recursive_check_for_a_user/cockroachdb/v1-10          8.25k          7.92k          
Services/recursive_check_for_a_user/mysql/v1-10                7.20k          6.80k          
Services/recursive_check_for_a_user/postgres/v1-10             7.42k          7.07k          
Services/recursive_check_for_a_user/spanner/v1-10              9.29k          8.89k          

Services/wide_groups_check_for_a_user/cockroachdb/v1-10         120k             8k          
Services/wide_groups_check_for_a_user/mysql/v1-10               119k             8k          
Services/wide_groups_check_for_a_user/postgres/v1-10            127k             8k          
Services/wide_groups_check_for_a_user/spanner/v1-10             170k            13k
```

The major difference is on the `wide_groups_check_for_a_user` test, where we see an approximately 3x performance boost due to the wide nature of that test

> **Note**
> The benchmarks were generated by running the various datastores in containers on a local M1, which likely explains why Postgres is so much faster: it is running under native `arm`. MySQL, CRDB, and Spanner are `amd64` images and therefore likely much slower than a real-world installation